==> 2_scripts/00_mobileID.ipynb <==
# generate info about mobile sites from mobile ID for both SANBS and Vitalant

import pandas as pd
import numpy as np
import datetime as dt
from datetime import timedelta

import matplotlib.pyplot as plt
import seaborn as sns

df=pd.read_csv("../1_data/private/SANBSdata.csv", error_bad_lines=False)
#df=pd.read_csv("../1_data/private/VitalantData.csv")

## for SANBS

# look at mobile ID to see the number of times a year it is visited 
#filter unique dates - some go for 2 or 3 consectutive days
df.drop(df[df['Visit_Date'].isna()].index, inplace=True) #remove any missing visit dates

df["Visit_Date"] = df["Visit_Date"].astype("datetime64") #convert to date time object
df.drop(df[(df['Fixed']== 'Lapsed Clinic') | (df['Fixed']=='Processing Lab')| (df['Fixed']=='Unallocated') 
           | (df["Fixed"]== 'Laboratorium') |(df["Fixed"]=='Doctor')].index, inplace=True)

df['Fixed_mobile'] = np.where(df['Fixed'] == 'Mobile Clinic', 0, 1) #mobile is 0, fixed is 1

df['Year']= df['Visit_Date'].dt.strftime('%Y')
df['Year']=df.Year.astype('int32')


df=df.loc[df.Fixed_mobile==0] #keep mobile sites
df_year_id=df.groupby(['Year', 'MobileID'])["Visit_Date"].apply(list)
df_year_id


df_mID = df_year_id.to_frame()
df_mID["Visit_Date"]=df_mID['Visit_Date'].apply(lambda x: np.unique(x)) #keep unique IDS
df_mID["Visit_Date"]=df_mID['Visit_Date'].apply(lambda x: sorted(x)) # sort
df_mID

# iterate through rows keep unique date that do not occur with one or two day of each other to determine frequency of drive
def filter_dates(date_list):
    filtered_dates = [date_list[0]]
    last_date = pd.to_datetime(date_list[0])

    for date_str in date_list[1:]:
        date = pd.to_datetime(date_str)
        duration = date - last_date
        days = duration.days
        if days >= 2:
            filtered_dates.append(date_str)
        last_date = date

    return filtered_dates

df_mID['Visit_Dates'] = df_mID['Visit_Date'].apply(filter_dates)


df_mID["Counts"]=df_mID['Visit_Dates'].apply(lambda x: len(x)) #count IDs


df_1= df_mID.explode('Visit_Date')
df_1.sort_values(by=['MobileID', 'Visit_Date'], axis = 0, inplace=True)
df_1=df_1.droplevel('Year')


df_2= df_mID.explode('Visit_Dates')
df_2.sort_values(by=['MobileID', 'Visit_Dates'], axis = 0, inplace=True)
df_2=df_2.droplevel('Year')
df_2.reset_index(inplace=True)


#count opp to donate ftom each date
def count_id_in_next_12_months(row):
    mask = (df_2['MobileID'] == row['MobileID']) & (df_2['Visit_Dates'] > row['Visit_Date']) & (df_2['Visit_Dates'] <= row['next_year'])
    return mask.sum()

df_1.reset_index(inplace=True)
df_1['next_year'] = df_1['Visit_Date'] + pd.DateOffset(months=12)
df_1['count_in_next_12_months'] = df_1.apply(count_id_in_next_12_months, axis=1)


df_1.to_csv(("../1_data/private/mID_table.csv"), index=False)


# For Vitalant
df=pd.read_csv("../1_data/private/VitalantData.csv")

df['MobileID']=df['COLLECTION_SITE_KEY']
df["Visit_Date"] = pd.to_datetime(df['DON_DATE_KEY'], format='%Y%m%d') #convert to datetime object


# look at mobile ID to see the number of times a year it is visited 
#filter unique dates - some go for 2 or 3 consectutive days
df.drop(df[df['Visit_Date'].isna()].index, inplace=True) #remove any missing visit dates
df["Visit_Date"] = df["Visit_Date"].astype("datetime64") #convert to date time object

df['Fixed_mobile'] = np.where(df["DRIVE_TYPE_FIXED_MOBILE"] == 'Fixed', 1, 0) #mobile is 0, fixed is 1

df['Year']= df['Visit_Date'].dt.strftime('%Y')
df['Year']=df.Year.astype('int32')

df=df.loc[df.Fixed_mobile==0] #keep mobile sites
df_year_id=df.groupby(['Year', 'MobileID'])["Visit_Date"].apply(list)
df_year_id


df_mID = df_year_id.to_frame()
df_mID["Visit_Date"]=df_mID['Visit_Date'].apply(lambda x: np.unique(x))
df_mID["Visit_Date"]=df_mID['Visit_Date'].apply(lambda x: sorted(x))
df_mID

df_mID['Visit_Dates'] = df_mID['Visit_Date'].apply(filter_dates)
df_mID["Counts"]=df_mID['Visit_Dates'].apply(lambda x: len(x))

df_1= df_mID.explode('Visit_Date')
df_1.sort_values(by=['MobileID', 'Visit_Date'], axis = 0, inplace=True)
df_1=df_1.droplevel('Year')

df_2= df_mID.explode('Visit_Dates')
df_2.sort_values(by=['MobileID', 'Visit_Dates'], axis = 0, inplace=True)
df_2=df_2.droplevel('Year')
df_2.reset_index(inplace=True)

df_1.reset_index(inplace=True)
df_1['next_year'] = df_1['Visit_Date'] + pd.DateOffset(months=12)
df_1['count_in_next_12_months'] = df_1.apply(count_id_in_next_12_months, axis=1)



df_1.to_csv(("../1_data/private/mID_table_V.csv"), index=False)
